# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de base en laravel",
        "description": "MDN",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://mdn-rdv.com/api"
        }
    ],
    "paths": {
...
        "/login": {
            "post": {
                "tags": [
                    "AUTHENTIFICATION"
                ],
                "summary": "Authentifie l'utilisateur avant de se connecter",
                "description": "",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "bgi@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "AUTHENTIFICATION"
                ],
                "summary": "Deconnexion de l'utilisateur",
                "description": "",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "schema": "Bearer",
                                        "type": "string",
                                        "example": "89er4186gjazuihhiZIOJreioiouEIOJIOZERF814879AE8FEA"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            },
            "0": []
        }
    }
}